requirements.txt
# PyQt6 for GUI
PyQt6==6.6.1

# Database
mysql-connector-python==8.2.0

# Data Science & ML
numpy==2.1.3
pandas==2.2.2
scikit-learn==1.6.1
imbalanced-learn==0.14.0

# ML Models
lightgbm==4.6.0
catboost==1.2.8
xgboost==2.0.3

# Plotting
matplotlib==3.9.2
seaborn==0.13.0

# Utilities
bcrypt==4.1.2
joblib==1.4.2
scipy==1.14.1
